requirements.txt
pymongo
requests
pyyaml
certifi
beautifulsoup4
markdownify
playwright
groq
flask
